[WIP] Windows: Added Windows support to `support/post-reviews.py`.
Windows: Added Windows support to \`support/post-reviews.py\`.

Early draft of the update to the Windows-compatible `post-reviews.py`. I'd like to get early feedback from people in case this is entirely the wrong direction.

At the outset, I will say that the "right way" to do this would be to refactor `cppylint.py` and `post-reviews.py` entirely so that they are more modular (for example, make `post-reviews.py` function-oriented, instead of a pile of statements and expressions, and `cpplint.py` should be a module proper). I didn't do this the "right way" because it is blocking for our Windows friends. If it's really important to do this the "right way", we may need to break it into a couple reviews.

Note that the most questionable part of this is the point at which we redirect `stderr` in `cpplint.py` to a string capture it for use in `post-reviews.py`. This is certainly the quickest way to do it, but it is not the best -- the best way would be to refactor `cpplint.py` to be a module, at least, and to handle error logging in a more pluggable way. We aimed for the shortest diff because there are no tests for any of these scripts, and to be honest, I was afraid of changing them without tests.

Review: https://reviews.apache.org/r/40340