Windows: Removed the `--switch_user` flag in Windows.
The Mesos agent provides a flag, `--[no-]switch_user`, which instructs
the agent to run tasks as the user who submitted the task instead of the
user that is running the agent process itself.

On POSIX, this requires the usual suspects (`su`, `chown`, and `chroot`,
for example to do things like make sure the correct user owns the
executor directory, and things like that.

On Windows, many of these operations have very different semantics. To
use one example, to `chown` on Windows, you normally have to give
*permission* for something to `chown` an object, and then that thing
actually has to use that permission to `chown` it, on its own.

Clearly a semantic mismatch of this magnitude will require some work to
formally bridge, and the result is that we will simply disable the
functionality entirely on the Windows agent until we are prepared to go
back and support it formally.

Review: https://reviews.apache.org/r/42035